+++ packages/xchain-bitcoincash/src/types/node-api-types.ts
@@ -1,5 +1,0 @@
-export type TxBroadcastResponse = {
-  id: string
-  result: string
-  error: string | null
-}